+++ Boss.h
@@ -16,7 +16,8 @@ public:
     Boss(const sf::Texture &bossTexture, const sf::Vector2i &spawnTile, const sf::Vector2i &tileSize,
          const sf::Vector2i &rectSkin, sf::Vector2f damageHit, const std::vector<std::vector<Node>> &nodeMap,
          float hitProbability, float hp = 50, sf::Vector2f scaleSize = {3, 3}, bool animated = true,
-         float speed = 200.f, unsigned int points = 100, unsigned int coins = 12, int armor = 10, bool bubble = false);
+         float speed = 200.f, unsigned int points = 100, unsigned int coins = 12, float armor = 10,
+         bool bubble = false);
 };
