+++ Zombie.h
@@ -13,7 +13,7 @@ public:
     Zombie(const sf::Texture &zombieTexture, const sf::Vector2i &spawnTile, const sf::Vector2i &tileSize,
            const sf::Vector2i &rectSkin, sf::Vector2f damageHit, const std::vector<std::vector<Node>> &nodeMap,
            float hitProbability, bool animated = true, float hp = 15, float speed = 230.f, unsigned int points = 5,
-           unsigned int coins = 1, int armor = 0, bool bubble = false);
+           unsigned int coins = 1, float armor = 0, bool bubble = false);
 
 };
